Backend/requirements_functions.txt
azure-storage-blob==12.26.0
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.3.0
tensorflow==2.13.0
scipy==1.11.3
opencv-python-headless==4.8.1.78
joblib==1.3.2
requests==2.31.0
